+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Changes to Calva Power Tools
 ## [Unreleased]
 
 - [Add performance tools](https://github.com/BetterThanTomorrow/calva-power-tools/issues/21)
+- [Add snitching for `let` and `fn`](https://github.com/BetterThanTomorrow/calva-power-tools/issues/24)
+- [Add snitching for current form](https://github.com/BetterThanTomorrow/calva-power-tools/issues/25)
 
 ## [v0.0.6] - 2025-05-06